+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+  - PODAAC-5424
+    - Update concise test in uat to test only POCLOUD collections
 ### Changed 
 ### Deprecated 
 ### Removed