+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 
   - Issue-68
     - Updated jupyter notebook
+    - Update notebook test to use python code directly instead of using jupyter notebook
+  - Updated python libraries
+
 ### Deprecated 
 ### Removed
 ### Fixed